web_server/requirements.txt
fastapi
uvicorn[standard]
gunicorn
motor
pydantic[email]
python-dotenv
sentence-transformers
httpx
python-docx
PyPDF2
python-multipart
PyJWT
cryptography
websockets
